rapid_haul_be/requirements.txt
Django==5.1.3
django-cors-headers==4.6.0
django-extensions==3.2.3
django-multiselectfield==0.1.13
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0